+++ src/gradedoc/add_template_comments.py
@@ -2,28 +2,16 @@
 
 from __future__ import annotations
 
-from typing import Optional
-
 import docxrev
 from win32com.client import constants
 
 from gradedoc import shared
-from gradedoc.shared import Path
 
 
-def add_template_comments(directory: Optional[Path] = None):
-    """Add template comments to all documents.
+def add_template_comments():
+    """Add template comments to all documents."""
 
-    Add template comments to all documents in `directory` as specified in
-    :py:func:`shared.get_paths`.
-
-    Parameters
-    ----------
-    directory
-        The directory containing the documents.
-    """
-
-    (paths, _) = shared.get_paths(directory)
+    (paths, _) = shared.get_paths()
 
     for path in paths:
         with docxrev.Document(path) as document: