+++ src/gradedoc/save_all.py
@@ -2,27 +2,15 @@
 
 from __future__ import annotations
 
-from typing import Optional
-
 import docxrev
 
 from gradedoc import shared
-from gradedoc.shared import Path
 
 
-def save_all(directory: Optional[Path] = None):
-    """Save all documents without closing them.
+def save_all():
+    """Save all documents without closing them."""
 
-    Save all documents in `directory` as specified in
-    :py:func:`shared.get_paths`.
-
-    Parameters
-    ----------
-    directory
-        The directory containing the documents.
-    """
-
-    (paths, _) = shared.get_paths(directory)
+    (paths, _) = shared.get_paths()
     for path in paths:
         document = docxrev.Document(path, save_on_exit=True, close_on_exit=False)
         with document: