+++ src/gradedoc/update_active_grade.py
@@ -2,43 +2,16 @@
 
 from __future__ import annotations
 
-import pathlib
-from typing import Optional
-from warnings import warn
-
 import docxrev
 
 from gradedoc import shared
-from gradedoc.shared import Path
 from gradedoc.update_grade import update_grade
 
 
-def update_active_grade(
-    directory: Optional[Path] = None, gradebook_path: Optional[Path] = None
-):
-    """Update the grade of the active document.
+def update_active_grade():
+    """Update the active document grade, insert common comments, update gradebook."""
 
-    Update the grade from the active document, and the gradebook. Ensure that the active
-    document and gradebook is as as specified in :py:func:`shared.get_paths`.
-
-    Parameters
-    ----------
-    directory
-        The directory where the active document should reside.
-    gradebook_name
-        The name of the gradebook (include ".csv"). Defaults to "grades.csv". Warn if
-        different than expected.
-    """
-
-    (paths, automatic_gradebook_path) = shared.get_paths(directory)
-
-    if gradebook_path is None:
-        gradebook_path = automatic_gradebook_path
-    else:
-        gradebook_path = pathlib.Path(gradebook_path)
-
-    if gradebook_path is not None and gradebook_path != automatic_gradebook_path:
-        warn("Supplied gradebook different than expected one.")
+    (paths, gradebook_path) = shared.get_paths()
 
     active_document = docxrev.get_active_document(save_on_exit=False)
     with active_document: